+++ lib/packmatic/source.ex
@@ -42,7 +42,9 @@ defmodule Packmatic.Source do
   The Source Entry `{:file, path}` is resolved during encoding:
 
       iex(1)> {:ok, file_path} = Briefly.create()
-      iex(2)> {:ok, state} = Packmatic.Source.build({:file, file_path})
+      iex(2)> {:ok, {module, state}} = Packmatic.Source.build({:file, file_path})
+      iex(3)> module
+      Packmatic.Source.File
       iex(3)> state.__struct__
       Packmatic.Source.File
 
@@ -68,13 +70,10 @@ defmodule Packmatic.Source do
   @typedoc """
   Represents the internal State for a resolved Source that is being read from.
 
-  Sources that hold state must use `defstruct` to define a struct, as the name of the struct is
-  used to refer them back to the Source module when reading data.
-
-  In case of a File source, the struct may hold the File Handle; in case of a URL source, it may
-  indirectly refer to the underlying network socket, etc.
+  In case of a File source, the state may be a struct which holds the File Handle; in case of a
+  URL source, it may be the underlying network socket, etc.
   """
-  @type t :: struct()
+  @type state :: term()
 
   @doc """
   Validates the given Initialisation Argument.
@@ -84,7 +83,7 @@ defmodule Packmatic.Source do
   @doc """
   Converts the Entry to a Source State.
   """
-  @callback init(term()) :: {:ok, t} | {:error, term()}
+  @callback init(term()) :: {:ok, state} | {:error, term()}
 
   @doc """
   Iterates the Source State.
@@ -93,7 +92,7 @@ defmodule Packmatic.Source do
   Source State can be returned with the data. Note that Sources that have returned `:eof` or
   `{:error, reason}` will not be touched again.
   """
-  @callback read(t) :: iodata() | {iodata(), t} | :eof | {:error, term()}
+  @callback read(state) :: iodata() | {iodata(), state} | :eof | {:error, term()}
 
   @typedoc """
   Represents an internal tuple that can be used to initialise a Source with `build/1`.
@@ -102,6 +101,9 @@ defmodule Packmatic.Source do
   work lazily, and other Sources may use this mechanism to delay opening of sockets or handles.
   """
   @type entry :: {name, init_arg}
+
+  @typedoc false
+  @opaque t :: {module(), state}
 
   @spec validate(entry) :: :ok | {:error, term()}
   @spec build(entry) :: {:ok, t} | {:error, term()}
@@ -129,8 +131,9 @@ defmodule Packmatic.Source do
   def build(entry)
 
   def build({name, init_arg}) do
-    with {:module, module} <- resolve(name) do
-      module.init(init_arg)
+    with {:module, module} <- resolve(name),
+         {:ok, state} <- module.init(init_arg) do
+      {:ok, {module, state}}
     end
   end
 
@@ -140,8 +143,13 @@ defmodule Packmatic.Source do
   Called by `Packmatic.Encoder`.
   """
   def read(state)
-  def read(%{__struct__: module} = state), do: module.read(state)
-  def read(_), do: {:error, :invalid_state}
+
+  def read({module, state}) do
+    case module.read(state) do
+      {data, state} when is_binary(data) or is_list(data) -> {data, {module, state}}
+      result -> result
+    end
+  end
 
   defp resolve(:file), do: {:module, __MODULE__.File}
   defp resolve(:url), do: {:module, __MODULE__.URL}